BJ\HW_9\actual version

+++ src/webapp/storage/AbstractArrayStorage.java
@@ -31,7 +31,7 @@ public abstract class AbstractArrayStorage extends AbstractStorage<Integer> {
         if (countResume >= storage.length) {
             throw new StorageException("Storage overflow ", r.getUuid());
         } else {
-            insertResume(r, (Integer) index);
+            insertResume(r, index);
             countResume++;
         }
     }
